refactor: consolidate background context into window context and update related components
