+++ src/App.tsx
@@ -1,20 +1,17 @@
 import { AssetProvider } from './context/AssetProvider'
-import { BackgroundProvider } from './context/BackgroundProvider'
 import SoundProvider from './context/SoundProvider'
 import WindowProvider from './context/WindowProvider'
 import Routes from './routes'
 
 function App() {
   return (
-    <WindowProvider>
-      <AssetProvider>
-        <BackgroundProvider>
-          <SoundProvider>
-            <Routes />
-          </SoundProvider>
-        </BackgroundProvider>
-      </AssetProvider>
-    </WindowProvider>
+    <AssetProvider>
+      <WindowProvider>
+        <SoundProvider>
+          <Routes />
+        </SoundProvider>
+      </WindowProvider>
+    </AssetProvider>
   )
 }
