+++ src/context/SoundProvider.tsx
@@ -1,8 +1,8 @@
 import { useCallback, useEffect, useRef, useState, type ReactNode } from 'react'
 
 import { useAssets } from './AssetContext'
-import { useBackground } from './BackgroundContext'
 import { SoundContext } from './SoundContext'
+import { useWindow } from './WindowContext'
 
 export default function SoundProvider({ children }: { children: ReactNode }) {
   const [isEffectMuted, setIsEffectMuted] = useState(
@@ -24,7 +24,7 @@ export default function SoundProvider({ children }: { children: ReactNode }) {
   }, [musicVolume, playingMusic])
 
   const { sounds } = useAssets()
-  const { interacted } = useBackground()
+  const { interacted } = useWindow()
 
   function playMusic(key: string) {
     if (!interacted) return
