+++ src/layouts/EntryLayout.tsx
@@ -4,13 +4,11 @@ import { Outlet } from 'react-router-dom'
 import logo from '@/assets/images/logo/logo.png'
 import Img from '@/components/common/Img'
 import SettingModalButton from '@/components/common/Settings'
-import { useBackground } from '@/context/BackgroundContext'
 import { useSound } from '@/context/SoundContext'
 import { useWindow } from '@/context/WindowContext'
 
 export default function EntryLayout() {
-  const { setBackgroundImage } = useBackground()
-  const { setIsCompact } = useWindow()
+  const { setBackgroundImage, setIsCompact } = useWindow()
   const { playMusic } = useSound()
 
   useEffect(() => {
