+++ src/layouts/GameRoomLayout.tsx
@@ -2,7 +2,6 @@ import { useEffect } from 'react'
 import { Outlet } from 'react-router-dom'
 
 import { useLeaveRoom } from '@/apis/room'
-import { useBackground } from '@/context/BackgroundContext'
 import { useWindow } from '@/context/WindowContext'
 import { useRoomStore } from '@/store/roomStore'
 
@@ -10,8 +9,7 @@ export default function GameRoomLayout() {
   const { mutate: leaveRoom } = useLeaveRoom()
   const { roomCode, flush } = useRoomStore()
 
-  const { setBackgroundImage } = useBackground()
-  const { setIsCompact } = useWindow()
+  const { setIsCompact, setBackgroundImage } = useWindow()
 
   useEffect(() => {
     setBackgroundImage('room')
